RavenDB-20628 - Move the deserialization from JsonDeserializationCluster to ClusterTransactionCommand (for FromRemote method)
