+++ src/Raven.Server/ServerWide/JsonDeserializationCluster.cs
@@ -155,8 +155,6 @@ namespace Raven.Server.ServerWide
         public static readonly Func<BlittableJsonReaderObject, OrchestratorConfiguration> OrchestratorConfiguration = GenerateJsonDeserializationRoutine<OrchestratorConfiguration>();
         public static readonly Func<BlittableJsonReaderObject, ShardingConfiguration> ShardingConfiguration = GenerateJsonDeserializationRoutine<ShardingConfiguration>();
 
-        public static Func<BlittableJsonReaderObject, ClusterTransactionResult> ClusterTransactionResult = GenerateJsonDeserializationRoutine<ClusterTransactionResult>();
-
         public static readonly Dictionary<string, Func<BlittableJsonReaderObject, CommandBase>> Commands = new Dictionary<string, Func<BlittableJsonReaderObject, CommandBase>>
         {
             [nameof(UnregisterReplicationHubAccessCommand)] = GenerateJsonDeserializationRoutine<UnregisterReplicationHubAccessCommand>(),
